fix error at ex007 (change int to float)

+++ workouts/ex007.py
@@ -1,6 +1,6 @@
 # Desenvolva um programa que leia duas notas de um aluno, calcule e motre a sua média
-note1 = int(input('Digite a primeira nota '))
-note2 = int(input('Digite a segunda nota '))
+note1 = float(input('Digite a primeira nota '))
+note2 = float(input('Digite a segunda nota '))
 
 media = (note1 + note2) / 2
 print(f'A média desse aluno é {media}')
